% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jffs-package.R
\docType{package}
\name{jffs-package}
\alias{jffs}
\alias{jffs-package}
\title{jffs: Mamíferos de la isla Juan Fernández}
\description{
Análisis y datos recopilados respecto a los mamíferos de la isla Juan Ferández.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mammal-proj/jffs}
  \item \url{https://mammal-proj.github.io/jffs/}
  \item Report bugs at \url{https://github.com/mammal-proj/jffs/issues}
}

}
\author{
\strong{Maintainer}: Matías Castillo-Aguilar \email{m99castillo@gmail.com} (\href{https://orcid.org/0000-0001-7291-247X}{ORCID})

Authors:
\itemize{
  \item Renato Borras-Chavez \email{rborras@gmail.com} (\href{https://orcid.org/0000-0002-6415-2121}{ORCID})
  \item Carla Rivera Rebella \email{rodofitas@gmail.com} (\href{https://orcid.org/0000-0001-8163-3325}{ORCID})
}

}
\keyword{internal}
